Gcc sanitizers 41

This branch (in its current stage) enables the UndefinedBehaviorSanitzer in debug builds (OPT=dbg) with gcc >= 4.9 on Linux.
This leads to some slow down in runtime (~25%, see res files in #41), but also enables checks that we usually don't do, especially the division-by-zero seems to be useful to me (it pointed out a few places where this happened and only for some of them this was allowed behavior).
If building with SANITIZE=full, then also memory checkers and thread checkers (if PARASCIP=true) are enabled.

I have taken care of the warnings/errors reported by the sanitziers (OPT=dbg SANITIZE=full PARASCIP=false) on MMM and MINLP_0s_1m with TIME=300.

See merge request !35

+++ src/scip/presol_gateextraction.c
@@ -183,7 +183,7 @@ SCIP_DECL_HASHKEYVAL(hashdataKeyValCons)
    assert(hashdata->nvars == 2);
 
    /* if we have only two variables we store at each 16 bits of the hash value the index of a variable */
-   hashval = (SCIPvarGetIndex(hashdata->vars[1]) << 16) + SCIPvarGetIndex(hashdata->vars[0]); /*lint !e701*/
+   hashval = ((unsigned int)SCIPvarGetIndex(hashdata->vars[1]) << 16) + SCIPvarGetIndex(hashdata->vars[0]); /*lint !e701*/
 
    return hashval;
 }
@@ -253,7 +253,7 @@ SCIP_DECL_HASHKEYVAL(setppcHashdataKeyValCons)
     * all variables these bitfields are combined by an or operation to get a good hashvalue for distinguishing the data
     */
    for( v = 1; v >= 0; --v )
-      hashval |= (1 << (SCIPvarGetIndex(hashdata->vars[v]) % 32)); /*lint !e701*/
+      hashval |= (1U << (SCIPvarGetIndex(hashdata->vars[v]) % 32)); /*lint !e701*/
 
    return hashval;
 }
